Altered to read from file instead of STDIN
Parser now properly parses the Processes given in the input file.
This drastically changed a lot of my input logic, so I created a
ThreadTimeline class to encapsulate the concept.  Essentially, this
allows me to ask the timeline if there are any new processes at a
given time.  While this COULD be used for future knowledge, that will
be taken care of inside the Scheduler logic, and we'll only use
the current time as determined by the System::CLOCK