Intermediate submission ready.
I think...

Added terminal-table gem (vendorized) to print pretty output.

Also made changes to work on Ruby 1.8.7-p249, which is what is
on the Alamode machines.  I assume that that is what the src will
be graded on.  If not, it should also run on 1.9.2 and 1.9.3.
Most of these changes involved reverting the new hash sytax
  {out => STDOUT}  becomes  {out: STDOUT}
The only other change I found was that the Array#rotate! method did
not exist in 1.8.7, so I had to remove it from my logic.